Flatten notifications and push subscriptions into Hunter
