+++ lib/hunter/web_push_subscription.ex
@@ -17,7 +17,6 @@ defmodule Hunter.WebPushSubscription do
       `admin.report`)
 
   """
-  alias Hunter.Api.HTTPClient
 
   @type t :: %__MODULE__{
           id: String.t(),
@@ -29,65 +28,4 @@ defmodule Hunter.WebPushSubscription do
 
   @derive [Poison.Encoder]
   defstruct [:id, :endpoint, :standard, :server_key, :alerts]
-
-  @doc """
-  Subscribe to Web Push notifications; each access token can have exactly
-  one subscription, and creating a new one replaces it
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `subscription` - map with `endpoint` (where push alerts will be sent),
-      `keys` (map with the `p256dh` public key and `auth` secret) and
-      optionally `standard` (use standardized web push)
-    * `data` - optional map with `alerts` (map of notification types to
-      booleans) and `policy` (`all`, `followed`, `follower` or `none`)
-
-  """
-  @spec create_push_subscription(Hunter.Client.t(), map, map) :: Hunter.WebPushSubscription.t()
-  def create_push_subscription(conn, subscription, data \\ %{}) do
-    HTTPClient.create_push_subscription(conn, subscription, data)
-  end
-
-  @doc """
-  Retrieve the Web Push subscription tied to the access token
-
-  ## Parameters
-
-    * `conn` - connection credentials
-
-  """
-  @spec push_subscription(Hunter.Client.t()) :: Hunter.WebPushSubscription.t()
-  def push_subscription(conn) do
-    HTTPClient.push_subscription(conn)
-  end
-
-  @doc """
-  Update the `data` portion of the Web Push subscription (which alerts to
-  receive and the delivery policy)
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `data` - map with `alerts` (map of notification types to booleans)
-      and/or `policy` (`all`, `followed`, `follower` or `none`)
-
-  """
-  @spec update_push_subscription(Hunter.Client.t(), map) :: Hunter.WebPushSubscription.t()
-  def update_push_subscription(conn, data) do
-    HTTPClient.update_push_subscription(conn, data)
-  end
-
-  @doc """
-  Remove the Web Push subscription tied to the access token
-
-  ## Parameters
-
-    * `conn` - connection credentials
-
-  """
-  @spec delete_push_subscription(Hunter.Client.t()) :: boolean
-  def delete_push_subscription(conn) do
-    HTTPClient.delete_push_subscription(conn)
-  end
 end
